+++ zakurad/src/components/mempool/tests/vector.rs
@@ -2355,6 +2355,7 @@ async fn setup_with_mempool_config_and_misbehavior_sender(
     let (sync_status, recent_syncs) = SyncStatus::new();
     let (mempool, mempool_transaction_subscriber) = Mempool::new(
         &mempool_config,
+        false,
         Buffer::new(BoxService::new(peer_set.clone()), 1),
         state_service.clone(),
         Buffer::new(BoxService::new(tx_verifier.clone()), 1),
@@ -2431,6 +2432,7 @@ async fn cancel_handles_drained_after_verification_timeout() {
         Timeout::new(peer_set, TRANSACTION_DOWNLOAD_TIMEOUT),
         Timeout::new(tx_verifier, TRANSACTION_VERIFY_TIMEOUT),
         state,
+        false,
         u64::MAX,
     ));